# api/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Find connected fullnodes peers and get their approximate location",
        "title": "Fullnodes Aggregator API",
        "contact": {},
        "version": "1.0"
    },
    "host": "map.alephium.notrustverify.ch",
    "basePath": "/",
    "paths": {
        "/fullnodes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fullnodes"
                ],
                "summary": "Get detected fullnodes peers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Last update",
                        "name": "lastUpdate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.FullnodeApi"
                            }
                        }
                    }
                }
            }
        },
        "/historic": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fullnodes"
                ],
                "summary": "Return number of nodes connected historically",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Limit number of historic data, default 50",
                        "name": "limt",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.NumNodesDb"
                            }
                        }
                    }
                }
            }
        },
        "/syncstatus": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fullnodes"
                ],
                "summary": "Get number of fullnodes synced and not synced",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.SyncCount"
                            }
                        }
                    }
                }
            }
        },
        "/versions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fullnodes"
                ],
                "summary": "Get version run by fullnodes peers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.ClientVersionCount"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ClientVersionCount": {
            "type": "object",
            "properties": {
                "clientVersion": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "main.FullnodeApi": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "clientVersion": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "isSynced": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "org": {
                    "type": "string"
                },
                "postal": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "main.NumNodesDb": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                }
            }
        },
        "main.SyncCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "isSynced": {
                    "type": "boolean"
                }
            }
        }
    }
}